chore: Rename avatar component to editableAvatar
