+++ app/javascript/dashboard/components-next/avatar/EditableAvatar.vue
